requirements.txt
aiogram==3.4.1
aiohttp==3.9.3
google-auth==2.29.0
gspread==5.12.4
